// types/interfaces/StylesInterface.ts
export interface Global {
  theme: {
    colors: {
      background: string;
      menu: string;
      header: string;
      body: string;
      footer: string;
      subfooter: string;
      bottom: string;
    };
    breakpoints: {
      xs: string;
      s: string;
      m: string;
      l: string;
      xl: string;
    };
  };
}